修老师：拿掉 4 处 int() 截断（把浮点价格截成整数）⇒ 天花板 21,213 → 20,383（−3.9%）
expand_rule_v10.py 的 4 处病灶：int(world.prices[...]) 把浮点价格截断
  int(1.60)=1 ⇒ 木价被看低 37%（S=1）；int(18.19)=18 ⇒ 只低 1%（S=10）
⇒ 老师说"排建筑预算"算错 ⇒ need=0 ⇒ 不排楼、木全卖掉、不买回。

改动 4 处（buy 的 px/q 余额封顶、预算 _budget、_wx、_need_gold）全改浮点；
cost_of 的 int() 经核是 no-op（成本表全是 int）。

实测（8 图 200 回合，同 seed）：
  修前 21,213（逐位复现旧指纹 24,261/24,866/…）
  修后 20,383（−3.9%；逐图 Δ −8.4/−2.8/−8.3/−1.2/−1.6/−1.7/−7.7/+1.1%）
⇒ 这个 artifact 值 3.9% 的老师消费，且方向是"老师靠它多花了钱"——
   修完老师不是变强，是变诚实。★验收线随之下移；一切"占教师 x%"的历史数须重算。

★ 这不是改平衡，是去 artifact ⇒ 无条件做；顺带它也让"缩放改变玩法"这件事小得多。
回滚点：git 历史 + ECS 侧 rl/runs/probe_ecs/expand_rule_v10.py.orig-20260914

+++ expand_rule_v10.py
@@ -262,8 +262,8 @@ def expand_rule_turn_v10(world, name: str, rng: random.Random | None = None,
 
     def buy(good, qty):
         """**市场调剂**：买多少由 `need` 定，不由余额定（不设 `reserve` 门槛）。"""
-        px = max(1, int(world.prices.get(good, 2)))
-        q = min(int(qty), max(0, int(R()["黄金"]) // px))
+        px = max(1.0, float(world.prices.get(good, 2)))
+        q = min(int(qty), max(0, int(int(R()["黄金"]) // px)))
         return q > 0 and do("buy", {"good": good, "qty": q}, world.buy, name, good, q)
 
     def sell(good, qty):
@@ -409,13 +409,13 @@ def expand_rule_turn_v10(world, name: str, rng: random.Random | None = None,
             continue
         _sur = int(R().get(_g, 0)) - need.get(_g, 0)
         if _sur > 0:
-            _budget += int(_sur * 0.9 * max(1, int(world.prices.get(_g, 2))))
-    _wx = max(1, int(world.prices.get("木头", 2)))
+            _budget += _sur * 0.9 * max(1.0, float(world.prices.get(_g, 2)))
+    _wx = max(1.0, float(world.prices.get("木头", 2)))
     planned: list[tuple[float, str, tuple]] = []
     for _pb, _bn, _p in roi:
         if len(planned) >= n_roi_slots:
             break
-        _need_gold = int(cost_of(_bn)) + int(BUILDINGS[_bn].get("wood", 0)) * _wx
+        _need_gold = cost_of(_bn) + BUILDINGS[_bn].get("wood", 0) * _wx
         if _need_gold > _budget:
             continue                     # 付不起就跳过（不中断：后面可能有更便宜的）
         _budget -= _need_gold
